+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2025-06-25 12:46:28 - Release v1.1.5
+
+### Changed Files:
+- `flexible-page-navigation.php`
+
 ## 2025-06-25 12:41:24 - 1.1.5
 
 ### Changed Files: